drivers: hv: vtl: Don't allow register ioctl on HCVMs
Usermode should not be issuing these ioctls for registers that are not
shared between VTLs, nor should we fallback to a hypercall on CVMs as
the hypervisor does not have access to register state.

On TDX this path should be restricted to just kernel owned registers
that are shared between the L1 and L2s. On SNP, this path should not
be called at all, as the VMSA is used to modify lower VMPL register state.

Prohibit using the ioctl as the malicious hypervisor while not
being able to access the register state might return anything.

Signed-off-by: Roman Kisel <romank@linux.microsoft.com>
---
* Conflicts:
	drivers/hv/mshv_vtl_main.c