drivers/hv/mshv_vtl: fix wait on sidecar CPU online
The sidecar driver's CPU online path tries to wait for the sidecar
kernel to finish deinitializing the CPU before it allows the kernel
to online it. However, the wait condition is wrong, causing the wait
to be ineffective when the sidecar kernel is slow and causing hangs
when the sidecar kernel is fast.

Fix the wait condition. Add a log entry as well for better diagnostics.

Signed-off-by: John Starks <jostarks@microsoft.com>