hcl-x64.config: Set CONFIG_ACPI=y
The wakeup mailbox that the virtual firmware implements to boot secondary
CPUs is defined in the ACPI specification (see version 6.6 section
5.2.12.19). The code in the kernel that makes use of the mailbox resides
in the x86 ACPI subsystem. CONFIG_ACPI needs to be set as 'y' to select it.

The option CONFIG_ACPI selects or enables many other configuration options
which in turn select more options that are not used with DeviceTree-based
firmware. Unselect all the options that have a menuconfig prompt.

The newly selected code remains dormant if the acpi=off is specified in the
kernel command line. The code that interacts with the mailbox remains
usable for DeviceTree platform firmware.

These are the options that are selected after running `make olddefconfig`
with this changeset:

 * CONFIG_ACPI_MADT_WAKEUP=y
   Enables the wakeup mailbox.

 * CONFIG_ACPI_LPIT
   Support for ACPI low-power idle table (not used with DeviceTree FW).

 * CONFIG_ACPI_LEGACY_TABLES_LOOKUP=y
   Tweaks ACPI root table lookup (not used with DeviceTree FW).

 * CONFIG_ARCH_MIGHT_HAVE_ACPI_PDC=y
 * CONFIG_ARCH_HAS_ACPI_TABLE_UPGRADE=y
   Architecture capabilities (unused).

 * CONFIG_ACPI_SYSTEM_POWER_STATES_SUPPORT=y
   Support for ACPI-based suspend/resume (not used with DeviceTree FW).

 * CONFIG_ACPI_HOTPLUG_IOAPIC=y
   Support for IO-APIC hotplug (not used with DeviceTree FW).

 * CONFIG_HAVE_ACPI_APEI=y
 * CONFIG_HAVE_ACPI_APEI_NMI=y
   Architecture capabilities (not used as CONFIG_ACPI_APEI=n).

 * CONFIG_PCI_LABEL=y
   ACPI-provided PCI naming facilities (not used with DeviceTree FW).

 * CONFIG_PNP=y
 * CONFIG_PNPACPI=y
   Support only. Does not add drivers.

 * CONFIG_FIRMWARE_TABLE=y
   Library for parsing ACPI tables.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>